[ci] review comments
